fix:refactor multitrack handling and enhance project media preparation
