+++ nodes/__init__.py
@@ -5,6 +5,7 @@ from .image import *
 from .ltx import *
 from .ltxv import *
 from .minimax import *
+from .project import *
 from .video import *
 from .subtitle import *
 from .wan import *
